+++ apps/storybook/src/TriggerDirective.stories.tsx
@@ -1,0 +1,40 @@
+import type { Meta, StoryObj } from '@storybook/preact'
+import { Campfire } from '@campfire/components'
+
+const meta: Meta = {
+  title: 'Campfire/Directives/Trigger'
+}
+
+export default meta
+
+/**
+ * Demonstrates the `trigger` directive toggling state and cleaning up on exit.
+ *
+ * @returns Campfire story showcasing the `trigger` directive.
+ */
+export const Basic: StoryObj = {
+  render: () => (
+    <>
+      <tw-storydata startnode='1' options='debug'>
+        <tw-passagedata pid='1' name='Start'>
+          {`
+:set[test=true]
+
+:::trigger{label="Click me"}
+  :set[test=false]
+:::
+
+:::if[!test]
+You clicked the button!
+:::
+
+:::onExit
+  :unset[test]
+:::
+`}
+        </tw-passagedata>
+      </tw-storydata>
+      <Campfire />
+    </>
+  )
+}
